aplicando boas praticas

+++ src/main/java/com/dev/produtosapi/controller/ProdutoController.java
@@ -1,7 +1,10 @@
 package com.dev.produtosapi.controller;
 
+import java.util.List;
 import java.util.UUID;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import com.dev.produtosapi.model.Produto;
@@ -16,27 +19,42 @@ public class ProdutoController {
         this.produtoRepository = produtoRepository;
     }
 
-    @PostMapping("/criar")
-    public Produto criarProduto(@RequestBody Produto produto) {
-        var id = UUID.randomUUID().toString();
-        produto.setId(id);
-        produtoRepository.save(produto);
+    @PostMapping
+    public ResponseEntity<Produto> criarProduto(@RequestBody Produto produto) {
+        produto.setId(UUID.randomUUID().toString());
+        var salvo = produtoRepository.save(produto);
         System.out.println("Produto criado com sucesso!" + produto.toString());
-        return produto;
-    }
-    @GetMapping("/listar")
-    public Produto listarProdutos() {
-        return produtoRepository.findAll().getFirst();
+        return ResponseEntity.status(HttpStatus.CREATED).body(salvo);
     }
 
-    @GetMapping("/obter/{id}")
-    public Produto obterProduto(@PathVariable String id) {
-        return produtoRepository.findById(id).orElse(null);
+    @GetMapping
+    public ResponseEntity<List<Produto>> listarProdutos(String nome) {
+        var produtos = produtoRepository.findAll();
+        return ResponseEntity.ok(produtos);
     }
 
-    @DeleteMapping("/deletar/{id}")
-    public void deletar(@PathVariable("id") String id ){
+    @GetMapping("/{id}")
+    public ResponseEntity<Produto> obterProduto(@PathVariable String id) {
+        return produtoRepository.findById(id).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deletar(@PathVariable("id") String id ) {
+        if (!produtoRepository.existsById(id)){
+            return ResponseEntity.notFound().build();
+        }
         produtoRepository.deleteById(id);
+        return ResponseEntity.noContent().build();
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Produto> atualizar(@PathVariable("id") String id, @RequestBody Produto produto){
+        if (!produtoRepository.existsById(id)){
+            return ResponseEntity.notFound().build();
+        }
+        produto.setId(id);
+        Produto atualizado = produtoRepository.save(produto);
+        return ResponseEntity.ok(atualizado);
     }
 
 }
